requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==22.0.0